pyproject.toml
[project]
name = "comet-toolbox"
version = "0.0.1"
authors = [
  {name="Micha Burkhardt", email="micha.burkhardt@uol.de"},
]
description = "Dynamic functional connectivity toolbox for multiverse analysis"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "numpy >=1.26.2",
  "scipy >=1.11.4",
  "nilearn >=0.10.2",
  "scikit-learn >=1.3.2",
  "statsmodels >=0.14.0",
  "matplotlib >=3.8.2",
  "nibabel >=5.1.0",
  "pandas >=2.1.3",
  "pycwt >=0.4.0b0",
  "bctpy >=0.6.1",
  "joblib >=1.3.2",
  "jinja2 >=3.1.2",
  "tqdm >=4.66.1",
  "numba >=0.58.1",
  "PyQt6-tools >=6.4.2",
  "qdarkstyle >=3.2.3"
]

[project.urls]
Homepage = "https://github.com/mibur1/dfc-multiverse"
Issues = "https://github.com/mibur1/dfc-multiverse/issues"

[project.scripts]
comet-gui = "comet.gui:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/comet"]
data-files = ["src/comet/resources", "src/comet/resources/img"]
